Removed link field from NoteEntity. Moved in utils directory: NetworkMonitor.kt and NoteAnalyticsService.kt.
