+++ app/src/main/java/com/svetlana/kuro/notesapp/domain/NoteEntity.kt
@@ -6,8 +6,7 @@ import androidx.room.PrimaryKey
 @Entity(tableName = "note_table")
 data class NoteEntity(
     var noteTitle: String?,
-    var noteDescription: String?,
-    var noteLink: String?
+    var noteDescription: String?
 ) {
     @PrimaryKey(autoGenerate = true)
     var noteId: Int = 0
