+++ app/src/main/res/values/strings.xml
@@ -3,7 +3,6 @@
 
     <string name="note_title">Note Title</string>
     <string name="note_description">Note Description</string>
-    <string name="edit_text_url">https://www.google.com/</string>
 
     <string name="add_note">Add Note</string>
     <string name="save_note">Save Note</string>
